requirements.txt
python-dotenv 
mss 
pillow 
pytesseract 
pyautogui 
requests 
keyboard
mcp[cli]
